+++ R/get_version_number.R
@@ -22,6 +22,7 @@
 #'
 #' @return A string containing the next (incremented) version number
 #' @export
+#' @family utility
 get_version_number <- function(path = ".") {
   assert_that(is_dir(path))
   if (interactive()) {
@@ -44,7 +45,7 @@ get_version_number <- function(path = ".") {
 
   # list all index.Rmd files
   indexpaths <- dir_ls(path = path, recurse = TRUE,
-                       regexp = "source/.*/index\\.Rmd")
+                       regexp = "source\\/s[fpioa]p\\/.+\\/index\\.Rmd")
 
   # read YAML front matter
   yamllists <- map(indexpaths, yaml_front_matter)
@@ -72,6 +73,9 @@ get_version_number <- function(path = ".") {
 #' @param versions a character vector with previously published version numbers
 #'
 #' @return A string containing the next (incremented) version number
+#' @export
+#' @family utility
+#'
 #'
 #' @importFrom assertthat assert_that
 #' @importFrom stringr str_detect