+++ R/render_protocol.R
@@ -16,6 +16,7 @@
 #' @importFrom fs path dir_exists dir_copy
 #'
 #' @export
+#' @family render
 #'
 #'
 #' @examples
@@ -38,16 +39,33 @@ render_protocol <- function(protocol_code = NULL,
       file.path(path_to_protocol, "css")
     )
   }
-
-  # render html
-  old_wd <- getwd()
-  setwd(dir = path_to_protocol)
+  on.exit(unlink(file.path(path_to_protocol, "css"),
+                 recursive = TRUE), add = TRUE)
+  # copy pandoc
+  if (!dir_exists(file.path(path_to_protocol, "pandoc"))) {
+    dir_copy(
+      system.file("pandoc", package = "protocolhelper"),
+      file.path(path_to_protocol, "pandoc")
+    )
+  }
+  on.exit(unlink(file.path(path_to_protocol, "pandoc"),
+                 recursive = TRUE), add = TRUE)
+  old_wd <- setwd(dir = path_to_protocol)
+  on.exit(setwd(old_wd), add = TRUE)
   suppressWarnings(
+    # render pdf
+    render_book(input = "index.Rmd",
+                output_dir = output_dir,
+                output_format = "bookdown::pdf_book",
+                envir = new.env(),
+                ...)
+    )
+  suppressWarnings(
+    # render html
     render_book(input = "index.Rmd",
               output_dir = output_dir,
               output_file = "index.html",
               envir = new.env(),
-              ...))
-  setwd(old_wd)
-
+              ...)
+  )
 }