+++ R/setup.R
@@ -1,0 +1,4 @@
+ignore_unused_imports <- function() {
+  slickR::slickR
+  reactable::reactable
+}